refactor(perf): split /students and /games pages into server shell + motion island
9th + 10th in the pattern series. Mirrors W17-W23 implementations.

- app/students/page.tsx: 843 LOC client → 71 LOC server (metadata + CollectionPage JSON-LD + BreadcrumbList)
- app/games/page.tsx: 885 LOC client → 69 LOC server (metadata + CollectionPage JSON-LD + BreadcrumbList)
- components/students/StudentsShell.tsx: motion + aurora + curated learning paths island
- components/games/GamesShell.tsx: motion + agentic pipeline + tech stack island

No copy changes. No FAQ extraction needed (neither page had a FAQ section).
TypeScript clean (npx tsc --noEmit).
